@@ -1,17 +1,16 @@
 apply plugin: 'com.android.library'
 
 dependencies {
-    compile "com.android.support:recyclerview-v7:23.1.1"
-    compile 'com.android.support:appcompat-v7:23.1.1'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
 }
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion '22.0.1'
+    compileSdkVersion 28
 
     defaultConfig {
-        minSdkVersion 7
-        targetSdkVersion 23
+        minSdkVersion 14
+        targetSdkVersion 28
     }
 }
 
