@@ -2,7 +2,7 @@ package com.shaubert.ui.adapters;
 
 import android.content.Context;
 import android.content.res.Resources;
-import android.support.v7.widget.ThemedSpinnerAdapter;
+import androidx.appcompat.widget.ThemedSpinnerAdapter;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
