@@ -1,6 +1,6 @@
 package com.shaubert.ui.adapters.common;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 public interface FragmentIndexResolver {
     int getFragmentIndex(Fragment fragment);
