@@ -1,7 +1,7 @@
 package com.shaubert.ui.adapters.common;
 
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.ViewTreeObserver;
 
 public class RecyclerViewHelper {
